refactor: move `project` into its own module
