+++ server/algorithms/src/lib.rs
@@ -2,6 +2,7 @@ use model;
 
 pub mod bootstrapping;
 pub mod clustering;
+mod project;
 pub mod routing;
 mod rtree;
 pub mod s2;
